feat: switch to ESM
BREAKING CHANGE: package now is `type:"module"` and distributed only
with ESM modules and ESNext lang level.
