+++ .editorconfig
@@ -1,16 +1,14 @@
 root = true
 
 [*]
-charset = utf-8
+indent_style = tab
+tab_width = 2
 end_of_line = lf
-indent_size = 2
-indent_style = space
+charset = utf-8
+trim_trailing_whitespace = true
 insert_final_newline = true
 max_line_length = 100
-tab_width = 2
-trim_trailing_whitespace = true
 
-[*.{md,mdx}]
-trim_trailing_whitespace = false
-indent_size = unset
-max_line_length = 100
+[*.yml]
+indent_style = space
+indent_size = 2
