--- .eslintrc.js
@@ -1,21 +1,0 @@
-module.exports = {
-  root: true,
-
-  ignorePatterns: ['cjs', 'esm', 'esnext', 'coverage', 'benchmark/dist'],
-
-  extends: ['@react-hookz/eslint-config/typescript'],
-
-  parserOptions: {
-    project: './tsconfig.eslint.json',
-    tsconfigRootDir: __dirname,
-  },
-
-  rules: {
-    '@typescript-eslint/no-use-before-define': 'off',
-    '@typescript-eslint/no-explicit-any': 'off',
-    '@typescript-eslint/no-unsafe-argument': 'off',
-    '@typescript-eslint/no-unsafe-call': 'off',
-    '@typescript-eslint/no-unsafe-member-access': 'off',
-    '@typescript-eslint/no-unsafe-assignment': 'off',
-  },
-};
