+++ .gitignore
@@ -1,8 +1,5 @@
 /.idea
 node_modules
 /coverage
-/cjs
-/esm
-/esnext
-/benchmark/dist
-yarn-error.log
+/src/**/*.js
+/src/**/*.d.ts
