+++ .npmignore
@@ -1,4 +1,3 @@
 *
-!/cjs
-!/esm
-!/esnext
+!/src/**/*.js
+!/src/**/*.d.ts
