+++ .prettierrc.cjs
@@ -1,0 +1,3 @@
+const config = require('@react-hookz/eslint-config/.prettierrc.cjs');
+
+module.exports = config;
