+++ benchmark/package.json
@@ -2,6 +2,7 @@
   "name": "benchamrk",
   "version": "1.0.0",
   "private": true,
+  "type": "module",
   "scripts": {
     "cleanup": "rimraf ./dist",
     "compile": "tsc",
@@ -10,17 +11,19 @@
     "benchmark": "yarn build && node ./dist/index.js"
   },
   "dependencies": {
-    "@types/nano-equal": "^2.0.0",
-    "@types/node": "^17.0.8",
+    "@react-hookz/deep-equal-local": "./..",
+    "@swc/core": "^1.3.107",
+    "@types/nano-equal": "^2.0.2",
+    "@types/node": "^20.11.16",
     "benchmark": "^2.1.4",
-    "dequal": "^2.0.2",
+    "browserify": "^17.0.0",
+    "dequal": "^2.0.3",
     "fast-deep-equal": "^3.1.3",
-    "react-fast-compare": "^3.2.0",
-    "ts-node": "^10.4.0",
-    "typescript": "^4.5.4",
-    "browserify": "^17.0.0"
+    "react-fast-compare": "^3.2.2",
+    "ts-node": "^10.9.2",
+    "typescript": "^5.3.3"
   },
   "devDependencies": {
-    "@types/benchmark": "^2.1.1"
+    "@types/benchmark": "^2.1.5"
   }
 }
