+++ benchmark/src/run.ts
@@ -1,37 +1,37 @@
-import { Suite } from 'benchmark';
+import benchmark from 'benchmark';
 
 const outputFn =
-  typeof document === 'undefined'
-    ? // eslint-disable-next-line no-console
-      console.log
-    : (text: string) => {
-        queueMicrotask(() => {
-          document.body.innerHTML += `${text.replace('\n', '<br/>')}<br/>`;
-        });
-      };
+	typeof document === 'undefined'
+		? console.log
+		: (text: string) => {
+				queueMicrotask(() => {
+					document.body.innerHTML += `${text.replace('\n', '<br/>')}<br/>`;
+				});
+			};
 
 export function run<T>(
-  testData: { name: string; data: T }[],
-  libraries: { [name: string]: (data: T) => void }
+	testData: Array<{ name: string; data: T }>,
+	libraries: Record<string, (data: T) => void>
 ) {
-  testData.forEach((test) => {
-    const suite = new Suite(test.name, {
-      onStart: () => {
-        outputFn(`\n# ${test.name}`);
-      },
-      onCycle: (ev: any) => {
-        outputFn(`  ${String(ev.target)}`);
-      },
-      onComplete: (ev: any) => {
-        // eslint-disable-next-line @typescript-eslint/restrict-template-expressions
-        outputFn(` Fastest is ${ev.currentTarget.filter('fastest').map('name')}`);
-      },
-    });
+	for (const test of testData) {
+		const suite = new benchmark.Suite(test.name, {
+			onStart() {
+				outputFn(`\n# ${test.name}`);
+			},
+			onCycle(ev: any) {
+				outputFn(`  ${String(ev.target)}`);
+			},
+			onComplete(ev: any) {
+				outputFn(` Fastest is ${ev.currentTarget.filter('fastest').map('name')}`);
+			},
+		});
 
-    Object.entries(libraries).forEach(([name, fn]) => {
-      suite.add(name, () => fn(test.data));
-    });
+		for (const [name, fn] of Object.entries(libraries)) {
+			suite.add(name, () => {
+				fn(test.data);
+			});
+		}
 
-    suite.run();
-  });
+		suite.run();
+	}
 }
