+++ commitlint.config.cjs
@@ -1,0 +1,6 @@
+module.exports = {
+	extends: ['@commitlint/config-conventional'],
+	rules: {
+		'footer-max-line-length': [1, 'always', 100],
+	},
+};
