--- commitlint.config.js
@@ -1,4 +1,0 @@
-module.exports = {
-  extends: ['@commitlint/config-conventional'],
-  rules: { 'footer-max-line-length': [1, 'always', 100] },
-};
