--- jest.config.js
@@ -1,14 +1,0 @@
-module.exports = {
-  preset: 'ts-jest',
-  testEnvironment: 'jsdom',
-  coverageDirectory: './coverage',
-  collectCoverage: true,
-  collectCoverageFrom: ['src/**/*.ts'],
-  globals: {
-    'ts-jest': {
-      diagnostics: {
-        ignoreCodes: [151_001],
-      },
-    },
-  },
-};
