+++ package.json
@@ -20,53 +20,55 @@
   "publishConfig": {
     "access": "public"
   },
-  "files": [
-    "cjs",
-    "esm",
-    "esnext"
-  ],
-  "main": "cjs/index.js",
-  "esnext": "esnext/index.js",
-  "types": "cjs/index.d.ts",
-  "module": "esm/index.js",
+  "type": "module",
+  "main": "./src/index.js",
+  "types": "./src/index.d.ts",
   "sideEffects": false,
+  "exports": {
+    ".": "./src/index.js"
+  },
+  "engines": {
+    "node": ">=18.0.0"
+  },
   "devDependencies": {
-    "@commitlint/cli": "^17.4.4",
-    "@commitlint/config-conventional": "^17.4.4",
-    "@commitlint/cz-commitlint": "^17.4.4",
-    "@jamesacarr/eslint-formatter-github-actions": "^0.2.0",
+    "@commitlint/cli": "^18.6.0",
+    "@commitlint/config-conventional": "^18.6.0",
+    "@commitlint/cz-commitlint": "^18.6.0",
     "@jamesacarr/jest-reporter-github-actions": "^0.0.4",
-    "@react-hookz/eslint-config": "^1.6.1",
+    "@react-hookz/eslint-config": "^2.1.5",
+    "@react-hookz/eslint-formatter-gha": "^1.0.1",
     "@semantic-release/changelog": "^6.0.2",
     "@semantic-release/git": "^10.0.1",
-    "@semantic-release/github": "^8.0.7",
-    "@types/jest": "^28.1.8",
+    "@semantic-release/github": "^9.2.6",
+    "@swc/core": "^1.3.107",
+    "@swc/jest": "^0.2.34",
+    "@types/jest": "^29.5.12",
     "commitizen": "^4.3.0",
-    "commitlint": "^17.5.0",
-    "concurrently": "^7.6.0",
-    "husky": "^8.0.3",
-    "jest": "^28.1.3",
+    "commitlint": "^18.6.0",
+    "eslint": "^8.56.0",
+    "eslint-plugin-prettier": "^5.1.3",
+    "husky": "^9.0.10",
+    "jest": "^29.7.0",
     "jest-environment-jsdom": "^29.5.0",
-    "lint-staged": "^13.2.0",
-    "semantic-release": "^19.0.5",
-    "ts-jest": "^28.0.8",
-    "ttypescript": "^1.5.15",
-    "typescript": "^4.8.4",
+    "lint-staged": "^15.2.1",
+    "prettier": "^3.2.5",
+    "rimraf": "^5.0.5",
+    "semantic-release": "^23.0.0",
+    "ts-jest": "^29.1.2",
+    "ts-node": "^10.9.2",
+    "typescript": "^5.3.3",
     "yarn": "^1.22.19"
   },
   "scripts": {
     "prepare": "husky install",
     "commit": "git-cz",
+    "build": "yarn build:clean && tsc --version && tsc",
+    "build:clean": "yarn rimraf -g ./src/**/*.js ./src/**/*.d.ts",
     "benchmark": "yarn build && cd ./benchmark && yarn && yarn benchmark",
-    "build": "yarn build:cleanup && concurrently yarn:build:cjs yarn:build:esm yarn:build:esnext --kill-others-on-fail",
-    "build:cleanup": "rimraf ./cjs ./esm ./esnext ./types",
-    "build:cjs": "ttsc -p ./tsconfig.build.json --module CommonJS --target ES5 --outDir ./cjs",
-    "build:esm": "ttsc -p ./tsconfig.build.json --module ES6 --target ES5 --outDir ./esm",
-    "build:esnext": "ttsc -p ./tsconfig.build.json --module ESNext --target ESNext --outDir ./esnext",
-    "lint": "eslint . --ext .js,.ts",
-    "lint:fix": "yarn lint --fix",
     "test": "jest",
-    "test:coverage": "yarn test --coverage"
+    "test:coverage": "yarn test --coverage",
+    "lint": "eslint .",
+    "lint:fix": "yarn lint --fix --fix-type problem,suggestion,layout"
   },
   "config": {
     "commitizen": {
@@ -74,7 +76,7 @@
     }
   },
   "lint-staged": {
-    "*.js": "yarn lint:fix"
+    "*.{js,jsx,ts,tsx,md,mdx}": "eslint --fix"
   },
   "release": {
     "plugins": [
