+++ src/comparators.ts
@@ -1,109 +1,108 @@
-/* eslint-disable @typescript-eslint/no-explicit-any,no-cond-assign,no-restricted-syntax,no-continue */
 type EqualFn = (a: any, b: any) => boolean;
 
 export const compareDates = (a: Date, b: Date): boolean => a.getTime() === b.getTime();
 
 export const compareRegexps = (a: RegExp, b: RegExp): boolean =>
-  a.source === b.source && a.flags === b.flags;
+	a.source === b.source && a.flags === b.flags;
 
 export const compareArrays = (a: any[], b: any[], equal: EqualFn): boolean => {
-  let l = a.length;
-  if (l !== b.length) return false;
+	let l = a.length;
+	if (l !== b.length) return false;
 
-  while (l-- && equal(a[l], b[l]));
+	while (l-- && equal(a[l], b[l]));
 
-  return l === -1;
+	return l === -1;
 };
 
 export const compareMaps = (a: Map<any, any>, b: Map<any, any>, equal: EqualFn): boolean => {
-  if (a.size !== b.size) return false;
-  const it = a.entries();
-  let i: any;
+	if (a.size !== b.size) return false;
+	const it = a.entries();
+	let i: any;
 
-  while (!(i = it.next()).done) {
-    if (!b.has(i.value[0]) || !equal(i.value[1], b.get(i.value[0]))) return false;
-  }
+	while (!(i = it.next()).done) {
+		if (!b.has(i.value[0]) || !equal(i.value[1], b.get(i.value[0]))) return false;
+	}
 
-  return true;
+	return true;
 };
 
 export const compareSets = (a: Set<any>, b: Set<any>): boolean => {
-  if (a.size !== b.size) return false;
-  const it = a.values();
-  let i: any;
+	if (a.size !== b.size) return false;
+	const it = a.values();
+	let i: any;
 
-  while (!(i = it.next()).done) {
-    if (!b.has(i.value)) return false;
-  }
+	while (!(i = it.next()).done) {
+		if (!b.has(i.value)) return false;
+	}
 
-  return true;
+	return true;
 };
 
 export const compareDataViews = (a: DataView, b: DataView): boolean => {
-  let l = a.byteLength;
-  if (l !== b.byteLength) return false;
+	let l = a.byteLength;
+	if (l !== b.byteLength) return false;
 
-  while (l-- && a.getInt8(l) === b.getInt8(l));
+	while (l-- && a.getInt8(l) === b.getInt8(l));
 
-  return l === -1;
+	return l === -1;
 };
 
 export const compareArrayBuffers = (a: ArrayLike<any>, b: ArrayLike<any>): boolean => {
-  let l = a.length;
-  if (l !== b.length) return false;
+	let l = a.length;
+	if (l !== b.length) return false;
 
-  while (l-- && a[l] === b[l]);
+	while (l-- && a[l] === b[l]);
 
-  return l === -1;
+	return l === -1;
 };
 
 const { hasOwnProperty } = Object.prototype;
 const oKeys = Object.keys;
 
 export const compareObjects = (
-  a: Record<any, any>,
-  b: Record<any, any>,
-  equal: EqualFn
+	a: Record<any, any>,
+	b: Record<any, any>,
+	equal: EqualFn
 ): boolean => {
-  let i;
-  let len = 0;
+	let i;
+	let length = 0;
 
-  for (i in a) {
-    if (hasOwnProperty.call(a, i)) {
-      len++;
+	for (i in a) {
+		if (hasOwnProperty.call(a, i)) {
+			length++;
 
-      if (!hasOwnProperty.call(b, i)) return false;
+			if (!hasOwnProperty.call(b, i)) return false;
 
-      if (!equal(a[i], b[i])) return false;
-    }
-  }
+			if (!equal(a[i], b[i])) return false;
+		}
+	}
 
-  return oKeys(b).length === len;
+	return oKeys(b).length === length;
 };
 
 export const compareObjectsReact = (
-  a: Record<any, any>,
-  b: Record<any, any>,
-  equal: EqualFn
+	a: Record<any, any>,
+	b: Record<any, any>,
+	equal: EqualFn
 ): boolean => {
-  let i;
-  let len = 0;
+	let i;
+	let length = 0;
 
-  for (i in a) {
-    if (hasOwnProperty.call(a, i)) {
-      len++;
-      if (a.$$typeof && (i === '_owner' || i === '__v' || i === '__o')) {
-        // in React and Preact these properties contain circular references
-        // .$$typeof is just reasonable marker of element
+	for (i in a) {
+		if (hasOwnProperty.call(a, i)) {
+			length++;
+			if (a.$$typeof && (i === '_owner' || i === '__v' || i === '__o')) {
+				// In React and Preact these properties contain circular references
+				// .$$typeof is just reasonable marker of element
 
-        continue;
-      }
+				continue;
+			}
 
-      if (!hasOwnProperty.call(b, i)) return false;
+			if (!hasOwnProperty.call(b, i)) return false;
 
-      if (!equal(a[i], b[i])) return false;
-    }
-  }
+			if (!equal(a[i], b[i])) return false;
+		}
+	}
 
-  return oKeys(b).length === len;
+	return oKeys(b).length === length;
 };
