--- tsconfig.build.json
@@ -1,20 +1,0 @@
-{
-  "extends": "./tsconfig.json",
-  "include": [
-    "src/index.ts"
-  ],
-  "compilerOptions": {
-    "noEmit": false,
-    "outDir": "./cjs",
-    "declaration": true,
-    "target": "ES5",
-    "module": "CommonJS",
-    "sourceMap": false,
-    "plugins": [
-      {
-        "transform": "./utility/ts-transformer-js-ext.ts",
-        "after": true
-      }
-    ]
-  }
-}
