+++ tsconfig.eslint.json
@@ -1,8 +1,22 @@
 {
-  "extends": "./tsconfig.json",
-  "include": [
-    "./.eslintrc.js",
-    "**/*.js",
-    "**/*.ts"
-  ]
+	"extends": "./tsconfig.json",
+	"compilerOptions": {
+		"noEmit": true
+	},
+	"exclude": [
+		"./dist",
+		"./node_modules",
+		"./coverage"
+	],
+	"include": [
+		".**/*.js",
+		".**/*.cjs",
+		"**/.*.js",
+		"**/.*.cjs",
+		"**/*.js",
+		"**/*.ts",
+		"**/*.tsx",
+		"**/*.md",
+		"**/*.mdx"
+	]
 }
